requirements.txt
python-redmine
pyyaml
python-dotenv
click
requests
